+++ dns.c
@@ -70,6 +70,7 @@ static void dns_query_complete(t_ref *ref, t_dns_elem *elem) {
       if (STR_EQ(orig->addr, elem->host.addr)) {
         UPD_NSTR(orig->name, elem->host.name, MAXHOSTNAME);
         if (hop->cached) hop->cached = false;
+        if (hop->cached_nl) hop->cached_nl = false;
         if (orig->name) stat_check_hostname_max(g_utf8_strlen(orig->name, MAXHOSTNAME));
         DNS_DEBUG("%s(%d,%d) addr=%s name=%s", __func__, hop->at, ndx, orig->addr, orig->name);
       } else LOG("dns(%s) origin is changed: %s", elem->host.addr, orig->addr);
@@ -152,6 +153,7 @@ void dns_lookup(t_hop *hop, int ndx) {
   if (cached) { // update with cached data and return
     UPD_NSTR(hop->host[ndx].name, cached->name, MAXHOSTNAME);
     if (hop->cached) hop->cached = false;
+    if (hop->cached_nl) hop->cached_nl = false;
     DNS_DEBUG("cache hit[%d,%d]: addr=%s -> name=%s", hop->at, ndx, addr, cached->name);
     return;
   }